Add computation of hashes for release files
This outputs a manifest for each file compatible with the format the
Installer uses so it can be easily copy-pasted to the configuration
file.

At the moment the installer does not accept https as a valid protocol so
that needs to be fixed before the GitHub URLs can be used to download
the files.