Open version_override.cmake in write mode instead of append mode.
If there's an existing version_override.cmake (which there is expected to be for a point-release, since it'll be based on the previous stable release), appending will cause there to be two sets of version information settings when we only want one. This shouldn't affect nightlies, but there's no reason for them to append either.